fix: feedback from PR
